move some  configuration stuff around
