+++ services/itops/src/nosql/Extensions.cs
@@ -5,6 +5,9 @@ namespace artiso.AdsdHotel.ITOps.NoSql
 {
     public static class Extensions
     {
+        public static bool HasData([NotNullWhen(true)] this string? s) =>
+            !string.IsNullOrWhiteSpace(s);
+            
         public static string NameForCollection(this Type type)
         {
             var entityTypeName = type.Name;
