+++ services/red/src/api/appsettings.Development.json
@@ -5,10 +5,5 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
-  },
-  "RabbitMqConfig": {
-    "Host": "localhost",
-    "User": "",
-    "Password": ""
   }
-}
+ }
