--- services/yellow/src/api/Extensions/Extensions.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Diagnostics.CodeAnalysis;
-using CaseExtensions;
-
-namespace artiso.AdsdHotel.Yellow.Api.Extensions
-{
-    public static class Extensions
-    {
-        public static bool HasData([NotNullWhen(true)] this string? s) =>
-            !string.IsNullOrWhiteSpace(s);
-        
-        public static string NameForCollection(this Type type)
-        {
-            var entityTypeName = type.Name;
-            var entityTypeFormattedName = entityTypeName.ToKebabCase();
-            return entityTypeFormattedName;
-        }
-    }
-}
