+++ services/yellow/src/api/appsettings.Development.json
@@ -5,10 +5,5 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
-  },
-  "RabbitMqConfig": {
-    "Host": "host=localhost",
-    "User": "",
-    "Password": ""
   }
 }
