+++ services/yellow/src/api/appsettings.json
@@ -6,18 +6,18 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "RabbitMqConfig": {
-    "Host": "host=localhost",
-    "User": "",
-    "Password": ""
-  },
   "MongoDbConfig": {
     "Host": "localhost",
-    "Port": 27016,
+    "Port": 27017,
     "User": "",
     "Password": "",
     "Database": "yellow",
     "Scheme": "mongodb"
+  },
+  "RabbitMqConfig": {
+    "Host": "localhost",
+    "User": "",
+    "Password": ""
   },
   "AllowedHosts": "*"
 }
